Add branch signal correction, using median beat. WIP: works in notebook, not in pipeline
